Add 'menu' and 'order' options
By requesting a 'GET' to the /menu endpoint, the
options of flavors and their price are shown.
By requesting a 'GET' to the /orders endpoint,
the order amount is shown.
By requesting a 'POST' to the /orders endpoint,
the order is placed. A fake credit card is used
to place a charge. 'stripe.com' is used for payment.